+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.2] — 2020-08-14
+
+-   Add embedding level to result of `Text::text_glyph_pos` (#17)
+-   Fix start offset for wrapped RTL text (#17)
+-   Fix `Text::line_index_nearest` for right-most position in line (#17)
+
 ## [0.1.1] — 2020-08-13
 
 -   `prepared::Text::positioned_glyphs` now takes an `FnMut` closure and